requirements.txt
python-dotenv
openai
python-docx
pydantic
PyPDF2
gradio

